apps/requirements.txt
faiss-cpu==1.10.0
langchain-community==0.3.21
langchain_google_genai==2.1.3
langchain==0.3.23
langchain_experimental==0.3.4
langgraph
pydantic
google-genai
pypdf
streamlit==1.44.1